gtest: internal: refactor common test functions out of sdpa
