+++ tests/gtests/internals/test_sdpa.cpp
@@ -18,6 +18,7 @@
 #include <gtest/gtest.h>
 
 #include "sdpa_internal.hpp"
+#include "test_utils.hpp"
 
 #include <oneapi/dnnl/dnnl.hpp>
 
@@ -25,15 +26,6 @@
 #include <random>
 
 using mdt = memory::data_type;
-
-enum class quantize_type {
-    no_quantization,
-    per_tensor,
-    per_tensor1,
-    per_tensor3,
-    per_token,
-    per_token_with_groups
-};
 
 enum class mask_type { no_mask, oneD, twoD, causal };
 
@@ -82,40 +74,6 @@ struct sdpa_tensors_t {
     int kq_mask, vs_mask;
     memory::dims kq_groups, vs_groups;
 };
-
-inline memory::dim product(const std::vector<int64_t> &dims) {
-    return dims.empty() ? 0
-                        : std::accumulate(dims.begin(), dims.end(),
-                                (memory::dim)1, std::multiplies<memory::dim>());
-}
-
-std::ostream &operator<<(std::ostream &ss, const quantize_type &qt) {
-    switch (qt) {
-        case quantize_type::no_quantization: ss << "no_quantization"; break;
-        case quantize_type::per_tensor: ss << "per_tensor"; break;
-        case quantize_type::per_tensor1: ss << "per_tensor1"; break;
-        case quantize_type::per_tensor3: ss << "per_tensor3"; break;
-        case quantize_type::per_token: ss << "per_token"; break;
-        case quantize_type::per_token_with_groups:
-            ss << "per_token_with_groups";
-            break;
-    }
-    return ss;
-}
-std::ostream &operator<<(std::ostream &ss, const memory::data_type &dt) {
-    switch (dt) {
-        case mdt::f32: ss << "f32"; break;
-        case mdt::s32: ss << "s32"; break;
-        case mdt::f16: ss << "f16"; break;
-        case mdt::bf16: ss << "bf16"; break;
-        case mdt::s8: ss << "s8"; break;
-        case mdt::u8: ss << "u8"; break;
-        case mdt::s4: ss << "s4"; break;
-        case mdt::u4: ss << "u4"; break;
-        default: ss << "na"; break;
-    }
-    return ss;
-}
 
 std::ostream &operator<<(std::ostream &ss, const sdpa_dims_t &p) {
     ss << "mb_" << p.mb;
@@ -225,169 +183,6 @@ using dnnl::softmax_forward;
         if (s_ != dnnl_success) COMPLAIN_DNNL_ERROR_AND_EXIT(#f, s_); \
     } while (0)
 
-/// Read from handle, write to memory
-/// This function is similar to the function found in write_to_dnnl_memory but this
-/// function has been expanded to perform an inline conversion from the source data
-/// type(handle) to the destination memory object. Currently, it only supports
-/// bf16, f16, f32, s32, u8/s8, u4/s4 data types for the dnnl::memory object
-/// and unsigned int and float for the handle.
-///
-/// The function first transfers the data to the handle's data type then
-/// uses reorder to perform the conversion to the destination data type.
-template <typename T>
-inline void write_to_dnnl_memory(const T *handle, dnnl::memory &mem) {
-    dnnl::engine eng = mem.get_engine();
-    size_t size = mem.get_desc().get_size();
-
-    if (!handle) throw std::runtime_error("handle is nullptr.");
-
-    dnnl::stream s(eng);
-    if (eng.get_kind() == dnnl::engine::kind::gpu) {
-        if (mem.get_desc().get_data_type() != dnnl_f32
-                && std::is_same<T, float>::value) {
-            memory mem_f32_mem(
-                    {mem.get_desc().get_dims(), memory::data_type::f32,
-                            mem.get_desc().get_strides()},
-                    eng);
-            write_to_dnnl_memory<float>((const float *)handle, mem_f32_mem);
-            dnnl::reorder(mem_f32_mem, mem).execute(s, mem_f32_mem, mem);
-            s.wait();
-        } else if (mem.get_desc().get_data_type() != dnnl_s32
-                && std::is_same<T, int>::value) {
-            memory mem_s32_mem(
-                    {mem.get_desc().get_dims(), memory::data_type::s32,
-                            mem.get_desc().get_strides()},
-                    eng);
-            write_to_dnnl_memory<int>((const int *)handle, mem_s32_mem);
-            dnnl::reorder(mem_s32_mem, mem).execute(s, mem_s32_mem, mem);
-            s.wait();
-        } else if ((mem.get_desc().get_data_type() == dnnl_u8
-                           || mem.get_desc().get_data_type() == dnnl_s8
-                           || mem.get_desc().get_data_type() == dnnl_s4
-                           || mem.get_desc().get_data_type() == dnnl_u4)
-                && std::is_same<T, unsigned>::value) {
-            memory mem_u32_mem(
-                    {mem.get_desc().get_dims(), memory::data_type::s32,
-                            mem.get_desc().get_strides()},
-                    eng);
-            write_to_dnnl_memory<unsigned>(
-                    (const unsigned *)handle, mem_u32_mem);
-            dnnl::reorder(mem_u32_mem, mem).execute(s, mem_u32_mem, mem);
-            s.wait();
-        } else if ((mem.get_desc().get_data_type() == dnnl_f32
-                           && std::is_same<T, float>::value)
-                || (mem.get_desc().get_data_type() == dnnl_s32
-                        && std::is_same<T, int>::value)) {
-            void *mapped_ptr = mem.map_data();
-            if (mapped_ptr) std::memcpy(mapped_ptr, handle, size);
-            mem.unmap_data(mapped_ptr);
-        } else {
-            // PC: this branch is identical to the one above
-            void *mapped_ptr = mem.map_data();
-            if (mapped_ptr) std::memcpy(mapped_ptr, handle, size);
-            mem.unmap_data(mapped_ptr);
-        }
-        return;
-    }
-
-    if (eng.get_kind() == dnnl::engine::kind::cpu) {
-        uint8_t *dst = static_cast<uint8_t *>(mem.get_data_handle());
-        if (!dst) throw std::runtime_error("get_data_handle returned nullptr.");
-        for (size_t i = 0; i < size; ++i)
-            dst[i] = ((uint8_t *)handle)[i];
-        return;
-    }
-
-    assert(!"not expected");
-}
-
-std::random_device &get_random_device() {
-    static std::random_device rd;
-    return rd;
-}
-
-std::mt19937 &get_generator() {
-    static std::mt19937 generator(get_random_device()());
-    return generator;
-}
-
-// this is changed from the fill_random() function in matmul_perf.cpp.
-void fill_random(std::vector<float> &out, const memory::desc &desc) {
-    static std::vector<float> random_data_f;
-    constexpr memory::dim nrand = 1037;
-
-    if (random_data_f.empty()) {
-        std::uniform_real_distribution<float> dist_f(-3.0f, 4.0f);
-
-        random_data_f.resize(nrand);
-        for (auto &d : random_data_f)
-            d = dist_f(get_generator());
-    }
-
-    auto elems = product(desc.get_dims());
-    for (memory::dim i = 0; i < elems; i += nrand) {
-        size_t chunk = std::min(nrand, elems - i);
-        std::memcpy(&out[i], random_data_f.data(), chunk * sizeof(float));
-    }
-}
-
-template <typename T>
-void fill_random_quantized(std::vector<T> &out, const memory::desc &desc,
-        bool is_unsigned = false) {
-    static std::vector<T> random_data_f;
-    static std::vector<T> random_data_u;
-    constexpr memory::dim nrand = 2049;
-
-    if (random_data_f.empty() || random_data_u.empty()) {
-        std::uniform_int_distribution<int> dist_f(-4, 4);
-        std::uniform_int_distribution<unsigned> dist_u(0, 6);
-
-        random_data_u.resize(nrand);
-        for (auto &d : random_data_u) {
-            d = dist_u(get_generator());
-        }
-        random_data_f.resize(nrand);
-        for (auto &d : random_data_f) {
-            d = dist_f(get_generator());
-        }
-    }
-
-    auto elems = product(desc.get_dims());
-    if (std::is_same<unsigned, T>::value || is_unsigned) {
-        for (memory::dim i = 0; i < elems; i += nrand) {
-            size_t chunk = std::min(nrand, elems - i);
-            std::memcpy(&out[i], random_data_u.data(), chunk * sizeof(T));
-        }
-    } else {
-        for (memory::dim i = 0; i < elems; i += nrand) {
-            size_t chunk = std::min(nrand, elems - i);
-            std::memcpy(&out[i], random_data_f.data(), chunk * sizeof(T));
-        }
-    }
-}
-
-void fill_random_scales(std::vector<float> &out, const memory::desc &desc) {
-    static std::vector<float> random_data_f;
-    constexpr memory::dim nrand = 1037;
-
-    if (random_data_f.empty()) {
-        std::uniform_int_distribution<int> dist_f(-16, 16);
-
-        random_data_f.resize(nrand);
-        for (auto &d : random_data_f) {
-            auto value = dist_f(get_generator()) * 0.125f;
-            if (value == 0.f) value = dist_f(get_generator()) * 0.125f;
-            d = value;
-        }
-    }
-
-    auto elems = product(desc.get_dims());
-    for (memory::dim i = 0; i < elems; i += nrand) {
-        size_t chunk = std::min(nrand, elems - i);
-        std::memcpy(&out[i], random_data_f.data(), chunk * sizeof(float));
-    }
-}
-
 // initialize the mask with first 3/4 elements with 0s and the last 1/4 elements
 // with -inf.
 void fill_mask(std::vector<float> &mask, const memory::desc &desc) {
@@ -427,285 +222,6 @@ void fill_causal_mask(std::vector<float> &mask, const memory::desc &desc) {
                 }
             }
         }
-    }
-}
-
-void print_mem(const dnnl::memory &mem, const std::string &name = "") {
-    auto eng = mem.get_engine();
-    dnnl::stream s(eng);
-    s.wait();
-    auto desc = mem.get_desc();
-    auto dims = desc.get_dims();
-    auto strides = desc.get_strides();
-    printf("%sbegin\n", name.c_str());
-    printf("dims   : %6ld %6ld %6ld %6ld\n", (long)dims[0], (long)dims[1],
-            (long)dims[2], (long)dims[3]);
-    printf("strides: %6ld %6ld %6ld %6ld\n", (long)strides[0], (long)strides[1],
-            (long)strides[2], (long)strides[3]);
-    if (mem.get_desc().get_data_type() == dnnl_bf16) { printf("bf16\n"); }
-    void *mapped_ptr_ = (void *)mem.map_data();
-    printf("        i:    ");
-    for (int i = 0; i < dims[3]; i++) {
-        switch ((int)desc.get_data_type()) {
-            case dnnl_u4:
-            case dnnl_s4: printf("%4d", i); break;
-            case dnnl_u8:
-            case dnnl_s8: printf("%4d", i); break;
-            case dnnl_f32:
-            case dnnl_bf16:
-            case dnnl_f16: printf("%9d", i); break;
-        }
-    }
-    printf("\n");
-
-    switch ((int)desc.get_data_type()) {
-        case dnnl_u4:
-        case dnnl_s4: {
-            char *mapped_ptr = (char *)mapped_ptr_;
-            for_(int l = 0; l < dims[0]; l++)
-            for_(int k = 0; k < dims[1]; k++)
-            for (int j = 0; j < dims[2]; j++) {
-                printf("(%2d, %2d, %3d):", l, k, j);
-                for (int i = 0; i < dims[3]; i++) {
-                    auto offset = l * strides[0] + k * strides[1]
-                            + j * strides[2] + i * strides[3];
-                    offset /= 2;
-                    bool is_odd = (strides[3] == 1) ? i % 2 : j % 2;
-                    if (is_odd) {
-                        int bits = (mapped_ptr[offset] & 0xf0) >> 4;
-                        if (desc.get_data_type() == dnnl_s4) {
-                            int sign = (bits & 0x08) ? -1 : 1;
-                            if (sign == -1) {
-                                bits = (bits & 0x07) - 8;
-                            } else {
-                                bits = (bits & 0x07);
-                            }
-                        }
-
-                        printf("%4d", bits);
-                    } else {
-                        int bits = (mapped_ptr[offset] & 0x0f);
-                        if (desc.get_data_type() == dnnl_s4) {
-                            int sign = (bits & 0x08) ? -1 : 1;
-                            if (sign == -1) {
-                                bits = (bits & 0x07) - 8;
-                            } else {
-                                bits = (bits & 0x07);
-                            }
-                        }
-
-                        printf("%4d", bits);
-                    }
-                }
-                printf("\n");
-            }
-        } break;
-
-        case dnnl_u8:
-        case dnnl_s8: {
-            char *mapped_ptr = (char *)mapped_ptr_;
-            for_(int l = 0; l < dims[0]; l++)
-            for_(int k = 0; k < dims[1]; k++)
-            for (int j = 0; j < dims[2]; j++) {
-                printf("(%2d, %2d, %3d): ", l, k, j);
-                for (int i = 0; i < dims[3]; i++) {
-                    printf("%4d",
-                            (mapped_ptr[l * strides[0] + k * strides[1]
-                                    + j * strides[2] + i * strides[3]]));
-                }
-                printf("\n");
-            }
-        } break;
-        case dnnl_bf16: {
-            using dnnl::impl::bfloat16_t;
-            bfloat16_t *mapped_ptr = (bfloat16_t *)mapped_ptr_;
-
-            for_(int l = 0; l < dims[0]; l++)
-            for_(int k = 0; k < dims[1]; k++)
-            for (int j = 0; j < dims[2]; j++) {
-                printf("(%2d, %2d, %3d):", l, k, j);
-                for (int i = 0; i < dims[3]; i++) {
-                    printf("%+9.3f",
-                            (float)(mapped_ptr[l * strides[0] + k * strides[1]
-                                    + j * strides[2] + i * strides[3]]));
-                }
-                printf("\n");
-            }
-        } break;
-        case dnnl_f16: {
-            using dnnl::impl::float16_t;
-            float16_t *mapped_ptr = (float16_t *)mapped_ptr_;
-
-            for_(int l = 0; l < dims[0]; l++)
-            for_(int k = 0; k < dims[1]; k++)
-            for (int j = 0; j < dims[2]; j++) {
-                printf("(%2d, %2d, %3d):", l, k, j);
-                for (int i = 0; i < dims[3]; i++) {
-                    printf("%+9.3f",
-                            (mapped_ptr[l * strides[0] + k * strides[1]
-                                    + j * strides[2] + i * strides[3]]
-                                            .f()));
-                }
-                printf("\n");
-            }
-        } break;
-        case dnnl_f32: {
-            using dnnl::impl::float16_t;
-            float *mapped_ptr = (float *)mapped_ptr_;
-
-            for_(int l = 0; l < dims[0]; l++)
-            for_(int k = 0; k < dims[1]; k++)
-            for (int j = 0; j < dims[2]; j++) {
-                printf("(%2d, %2d, %3d):", l, k, j);
-                for (int i = 0; i < dims[3]; i++) {
-                    printf("%+9.3f",
-                            (mapped_ptr[l * strides[0] + k * strides[1]
-                                    + j * strides[2] + i * strides[3]]));
-                }
-                printf("\n");
-            }
-        } break;
-        default: throw std::runtime_error("Not supported");
-    }
-    mem.unmap_data(mapped_ptr_);
-    printf("%send\n", name.c_str());
-}
-
-template <typename T>
-std::vector<float> dequantize(const std::vector<float> &input,
-        memory::desc &desc, memory::desc &scale_md,
-        const std::vector<T> &zero_points, const std::vector<float> &scales,
-        int group_size, quantize_type qtype, memory::dims groups,
-        int token_dim = -1) {
-    std::vector<float> out(input.size());
-    if (qtype == quantize_type::per_tensor
-            || qtype == quantize_type::per_tensor1
-            || qtype == quantize_type::per_tensor3) {
-        for (size_t i = 0; i < input.size(); i++)
-            out[i] = (input[i] - zero_points[i / group_size])
-                    * scales[i / group_size];
-    } else {
-        auto dims = desc.get_dims();
-        auto strides = desc.get_strides();
-
-        auto scales_dim = scale_md.get_dims();
-        auto zp_dim = scale_md.get_dims();
-        auto scales_strides = scale_md.get_strides();
-        auto zp_strides = scale_md.get_strides();
-
-        if (qtype == quantize_type::no_quantization) { groups = {1, 1}; }
-        if (qtype == quantize_type::per_token) {
-            if (token_dim == 0) {
-                groups = {dims[2], 1};
-            } else if (token_dim == 1) {
-                groups = {1, dims[3]};
-            }
-        }
-
-        int sg = 0;
-        int zg = 0;
-        int scale_offset = 0;
-        int zp_offset = 0;
-
-        for_(int l = 0; l < dims[0]; l++)
-        for_(int k = 0; k < dims[1]; k++)
-        for_(int j = 0; j < dims[2]; j++)
-        for (int i = 0; i < dims[3]; i++) {
-            if (groups[0] > 1) {
-                sg = groups[0];
-                scale_offset = l * scales_strides[0] + k * scales_strides[1]
-                        + j / sg * scales_strides[2] + i * scales_strides[3];
-            } else if (groups[1] > 1) {
-                sg = groups[1];
-                scale_offset = l * scales_strides[0] + k * scales_strides[1]
-                        + j * scales_strides[2] + i / sg * scales_strides[3];
-            }
-            if (groups[1] > 1) {
-                zg = groups[1];
-                zp_offset = l * zp_strides[0] + k * zp_strides[1]
-                        + j * zp_strides[2] + i / zg * zp_strides[3];
-            } else if (groups[0] > 1) {
-                zg = groups[0];
-                zp_offset = l * zp_strides[0] + k * zp_strides[1]
-                        + j / zg * zp_strides[2] + i;
-            }
-            int offset = l * strides[0] + k * strides[1] + j * strides[2]
-                    + i * strides[3];
-
-            out[offset] = (input[offset] - zero_points[zp_offset])
-                    * scales[scale_offset];
-        }
-
-        //int groups = zero_points.size();
-        //for (int g = 0; g < groups; g++) {
-        //    for (int i = g * group_size; i < g * group_size + group_size; i++) {
-        //        out[i] = (input[i] - zero_points[g]) * scales[g];
-        //        printf("out: %f = (input: %f - zero_point: %d) * scale: %f\n",
-        //                out[i], input[i], zero_points[g], scales[g]);
-        //    }
-        //}
-    }
-    return out;
-}
-
-void transpose(const dnnl::engine &eng, memory &out, memory &in) {
-    dnnl::stream s(eng);
-
-    void *ptr2 = out.map_data();
-    void *ptr1 = in.map_data();
-
-    std::memcpy(ptr2, ptr1, in.get_desc().get_size());
-    in.unmap_data(ptr1);
-    out.unmap_data(ptr2);
-}
-
-void transpose_strides(const dnnl::engine &eng, memory &out, memory &in) {
-    dnnl::stream s(eng);
-
-    if (out.get_desc().get_data_type() == mdt::u4
-            || out.get_desc().get_data_type() == mdt::s4) {
-        auto desc = in.get_desc();
-        auto dims = desc.get_dims();
-        auto strides = desc.get_strides();
-        auto strides_t = out.get_desc().get_strides();
-
-        char *mapped_ptr = (char *)in.map_data();
-        char *mapped_ptr_t = (char *)out.map_data();
-        for_(int l = 0; l < dims[0]; l++)
-        for_(int k = 0; k < dims[1]; k++)
-        for_(int j = 0; j < dims[2]; j++)
-        for (int i = 0; i < dims[3]; i++) {
-            int is_odd = i % 2;
-            int is_odd_t = j % 2;
-
-            auto offset = l * strides[0] + k * strides[1] + j * strides[2]
-                    + i * strides[3];
-            offset /= 2;
-
-            auto offset_t = l * strides_t[0] + k * strides_t[1]
-                    + j * strides_t[2] + i * strides_t[3];
-            offset_t /= 2;
-
-            auto &val = mapped_ptr[offset];
-            auto &val_t = mapped_ptr_t[offset_t];
-
-            char bits;
-            if (is_odd) {
-                bits = val & 0xf0;
-                bits >>= 4;
-            } else {
-                bits = val & 0x0f;
-            }
-            if (is_odd_t) {
-                val_t |= (bits << 4);
-            } else {
-                val_t |= bits;
-            }
-        }
-        in.unmap_data(mapped_ptr);
-        out.unmap_data(mapped_ptr_t);
-    } else {
-        dnnl::reorder(in, out).execute(s, in, out);
     }
 }
